requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
aiomysql==0.2.0
sqlalchemy[asyncio]==2.0.35
openpyxl==3.1.5
requests==2.32.3
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.5.2
python-dotenv==1.0.1
